requirements.txt
boto3==1.38.13
botocore==1.38.13
dnspython==2.7.0
jmespath==1.0.1
numpy==2.2.5
pandas==2.2.3
psycopg2-binary==2.9.10
pymongo==4.12.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
s3transfer==0.12.0
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
